Move HologramFormat code to a dedicated package and add comments to understand the code
